+++ scas/src/scas/residue/conversion/GaloisField.scala
@@ -3,6 +3,6 @@ package scas.residue.conversion
 import scas.polynomial.PolynomialOverField
 import scas.base.BigInteger
 
-class GaloisField[T, M](using PolynomialOverField[T, BigInteger, M]) extends Residue[T, BigInteger, M] {
+class GaloisField[T, M](using PolynomialOverField[T, Int, M]) extends Residue[T, Int, M] {
   given instance: GaloisField[T, M] = this
 }
